create icon svg objects [ci skip]
