+++ lib/parser.js
@@ -2,6 +2,7 @@ const jq = require('../lib/jsonquery.js')
 const hw = require('../lib/htmlWriter.js')
 const mj = require('../lib/modelicaToJSON.js')
 const ut = require('../lib/util.js')
+const svg = require('../lib/svg.js')
 
 var logger = require('winston')
 
@@ -208,8 +209,8 @@ function getJsons (moFile, parseMode, outputFormat) {
       }
       // Include .svg for icon layers in all classes and interfaces, .svg for diagram layer
       // in parsing class
-      ut.addIconSVG(data)
-      ut.addDiagramSVG(data)
+      svg.addIconSVG(data)
+      svg.addDiagramSVG(data)
     }
     // Order the input and output connectors.
     // This needs to be done on the whole data structure
